Fix reducers & actions,playing → active,melody→seq
Music buttons initial state had a «playing» field and, after some
tought, I believe it was an error. This state is more about
presentation. So it should be something talking about his visual
behaviour, so «active» looks better. And, with it, I needed to
correct the actions involved.

Same occurs with actions/state about «playingMelody». There is an
option about playing sound or not, so... if there isn't sound,
there isn't a melody. This is more about, again, visuals.
«sequence» seems more accurate here, because that's what is being
done here: replaying a sequence of moves/buttons, that can or not
have audio or visual cues (Just here raised an idea: a game mode
where there is only audio cues, not visual ones. A challenge for
the not blind people).