Update neutrino packages & entry point fix
Added a comment in the `src/index.js` to disable the ESLint rule
about not having `.js` files with JSX code. It should be an
`.jsx`, but the change doesn't seems easy in current Neutrino
version. In next version (5.0) it will be so easy.
As soon as `neutrino-preset-airbnb` get upgraded to it, I will
update this.