Transform startGame action creator into a thunk
It makes sense, because it does warn about the start of the game
and, in addition, also implies that a `NEW_TONE` needs to be
dispatched. Both could stay separated, but that would imply that
you need to remember to dispatch the second one every time you
dispatch the previous.