Add error animation & change handle of sequences
Till now, I «used» a `playSequence` boolean state to being aware of
a sequence in process by CPU. «used» because I wasn't using it at all.

Problem with this is that there are multiple sequences:
- A turn sequence, of music buttons being activated/played.
- An error. When an error happens, it plays something, and thats some
  kind of sequence.
- A (future) win sequence. To play something when the user wins the
  game.

So, there was 2 options:

1. To create more state keys, to be changed when these new sequences
   happen (to react when they change)
2. To modify the actual usage of `playSequence`

I went for the second. Generalize it, made it to cover all the
possible (CPU handled) sequences. When I did that, also discover that
the state key `allowUserInteraction` wasn't needed at all, because
the `playing` key was just enought to pointing that out. It was a
win-win situation, that spares some code into the reducers and
simplifies the state tree of the app.