Implement game mode «surprise»
Doing it was necessary to change a bit the styles structure,
as well ass adding a pair of new keys in the state, specifically
in the reducer `simon-buttons.js`. These states state (it's a pun)
the position of every music button of the Simon
(now their main CSS classes refer to them by their position, an
invariant value, in contrast with their assigned color) and the
actual distribution of colors.