Create lastEndScore redux state key
This simplifies the showing of final score in the game over screen
and the wind score (or something like that, still not implemented).
Also get rid of the bug about the app not clearing the
`currentScore` when changing from game over screen to main menu
screen. Error also originated because a double dispatching of
`END_GAME` and `RESET_GAME` actions in `SimonButtonGroup`.
Splitted the previous `leaveGame()` thunk in two functions, one that
even need access to actions or redux state, so, like it's so little,
end in the previously mentioned component.