build(build): change entry point of the app
From Neutrino V5 we can change the entry point of the app. So we change it now from the previously
forced `index.js` to `index.jsx`, because it contents `JSX` and also get rid of the previous ESLint
comment in this line, to allow usage of `JSX` in no `.jsx` files.