feat: add the «game won» behaviour
For it, generalized the previous `GameOverScreen` to a
`GameEndScreen`, that changes their title and buttons when the user
win/lose the game. Also, needed a container now
(for receiving the `gameWon`/`gameOver` state).
The component and the screen have the same filename, but like they
are placed in different folders, I guess is not so confussing
(I really couldn't figure out a better name to differentiate them :þ).

TODO: add animation to play when user wins the game. I'm waiting here
to some PR of `styled-jsx` to be finally merged.